Fix handling of "id=?" corner cases in rtree when the value on the RHS is a real value.
FossilOrigin-Name: 027e5336acc26f57f21df4980928731026c30cf88688fa0b66f13ffa0b5da3a0

+++ ext/rtree/rtree.c
@@ -1870,7 +1870,7 @@ static int rtreeFilter(
     i64 iNode = 0;
     int eType = sqlite3_value_numeric_type(argv[0]);
     if( eType==SQLITE_INTEGER
-     || (eType==SQLITE_FLOAT && sqlite3_value_double(argv[0])==iRowid)
+     || (eType==SQLITE_FLOAT && (i64)(sqlite3_value_double(argv[0]))==iRowid)
     ){
       rc = findLeafNode(pRtree, iRowid, &pLeaf, &iNode);
     }else{
